mega update

+++ spring/src/main/java/cse416/spring/database/CountyWriter.java
@@ -30,7 +30,7 @@ public class CountyWriter {
         StateName stateName = StateName.NORTH_CAROLINA;
 
         JSONObject jo = readJsonFile(countiesFilePath);
-        HashMap<Long, Precinct> allPrecincts = getAllPrecincts();
+        HashMap<Integer, Precinct> allPrecincts = getAllPrecincts();
         Iterator<String> keys = jo.keys();
 
         /* For each county */
